+++ smoke-tests/src/test/java/io/opentelemetry/smoketest/SdkDisabledSmokeTest.java
@@ -7,7 +7,6 @@ package io.opentelemetry.smoketest;
 
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.assertThat;
 
-import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import java.util.jar.Attributes;
 import java.util.jar.JarFile;
 import org.junit.jupiter.api.condition.DisabledIf;
@@ -19,23 +18,23 @@ import org.junit.jupiter.params.provider.ValueSource;
 class SdkDisabledSmokeTest {
 
   @RegisterExtension
-  static final SmokeTestTarget target =
-      SmokeTestTarget.springBoot("20211213.1570880324").env("OTEL_SDK_DISABLED", "true").build();
-
-  static final InstrumentationExtension testing = target.testing();
+  static final SmokeTestInstrumentationExtension testing =
+      SmokeTestInstrumentationExtension.springBoot("20211213.1570880324")
+          .env("OTEL_SDK_DISABLED", "true")
+          .build();
 
   @ParameterizedTest
   @ValueSource(ints = {8, 11, 17})
   void noopSdkSmokeTest(int jdk) throws Exception {
-    SmokeTestOutput output = target.start(jdk);
+    SmokeTestOutput output = testing.start(jdk);
     String currentAgentVersion =
-        new JarFile(target.getAgentPath())
+        new JarFile(testing.getAgentPath())
             .getManifest()
             .getMainAttributes()
             .get(Attributes.Name.IMPLEMENTATION_VERSION)
             .toString();
 
-    assertThat(target.client().get("/greeting").aggregate().join().contentUtf8()).isEqualTo("Hi!");
+    assertThat(testing.client().get("/greeting").aggregate().join().contentUtf8()).isEqualTo("Hi!");
     assertThat(testing.spans()).isEmpty();
     output.assertVersionLogged(currentAgentVersion);
     assertThat(testing.spans()).isEmpty();
